this is fully firebase read and write
