+++ app/src/main/java/com/example/allisonbolen/myapplication/MainActivity.java
@@ -11,16 +11,27 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.example.allisonbolen.myapplication.dummy.DummyContent;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.ChildEventListener;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
 
 
     private FirebaseAuth authUser;
+    public static DatabaseReference database;
+    public static List<DummyContent.Application_Information_Object> allApps;
 
 //    @Override
 //    public void onStart() {
@@ -43,12 +54,11 @@ public class MainActivity extends AppCompatActivity {
         EditText username = findViewById(R.id.username_editText);
         EditText password = findViewById(R.id.password_editText);
         InputMethodManager mgr = (InputMethodManager) getSystemService(Context.INPUT_METHOD_SERVICE);
+        allApps = new ArrayList<DummyContent.Application_Information_Object>();
 
         // ...
         // Initialize Firebase Auth
         authUser = FirebaseAuth.getInstance();
-
-
 
         // mode method
         signUp.setOnClickListener(v -> {
@@ -82,5 +92,50 @@ public class MainActivity extends AppCompatActivity {
 
     }
 
+    @Override
+    public void onResume(){
+        super.onResume();
+        allApps.clear();
+        database= FirebaseDatabase.getInstance().getReference("Cards");
+        database.addChildEventListener (chEvListener);
 
+    }
+
+    private ChildEventListener chEvListener = new ChildEventListener() {
+        @Override
+        public void onChildAdded(DataSnapshot dataSnapshot, String s) {
+            DummyContent.Application_Information_Object entry =
+                    (DummyContent.Application_Information_Object) dataSnapshot.getValue(DummyContent.Application_Information_Object.class);
+            entry._key = dataSnapshot.getKey();
+            allApps.add(entry);
+        }
+
+        @Override
+        public void onChildChanged(DataSnapshot dataSnapshot, String s) {
+        }
+
+        @Override
+        public void onChildRemoved(DataSnapshot dataSnapshot) {
+            DummyContent.Application_Information_Object entry =
+                    (DummyContent.Application_Information_Object) dataSnapshot.getValue(DummyContent.Application_Information_Object.class);
+            entry._key = dataSnapshot.getKey();
+            List<DummyContent.Application_Information_Object> newApp = new ArrayList<DummyContent.Application_Information_Object>();
+            for (DummyContent.Application_Information_Object t : allApps) {
+                if (!t._key.equals(dataSnapshot.getKey())) {
+                    newApp.add(t);
+                }
+            }
+            allApps = newApp;
+        }
+
+        @Override
+        public void onChildMoved(DataSnapshot dataSnapshot, String s) {
+
+        }
+
+        @Override
+        public void onCancelled(DatabaseError databaseError) {
+
+        }
+    };
 }
